requirements.txt
openai>=1.0.0
pillow>=10.0.0
transformers>=4.30.0
tiktoken>=0.5.0


